+++ Amazon.Pay.API.SDK/WebStore/Types/AddressDetails.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace Amazon.Pay.API.WebStore.Types
+{
+    public class AddressDetails : AddressBase
+    {
+        /// <summary>
+        /// Distrit or County of the address.
+        /// </summary>
+        [JsonProperty(PropertyName = "districtOrCounty")]
+        public string DistrictOrCounty { get; set; }
+    }
+}