+++ Amazon.Pay.API.SDK/WebStore/Types/CheckoutMode.cs
@@ -1,0 +1,14 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace Amazon.Pay.API.WebStore.Types
+{
+    /// <summary>
+    /// Specify whether the buyer will return to your website to review their order before completing checkout.
+    /// </summary>
+    [JsonConverter(typeof(StringEnumConverter))]
+    public enum CheckoutMode
+    {
+        ProcessOrder
+    }
+}